core: device-factory: implement match_connection()
Make it possible to register different factories for the same setting
type, and add a match_connection() method to let each factory decide
if it's capable of handling a connection.

This will be used to decide whether a PPPoE connection must be handled
through the legacy Ethernet factory or through the new PPP factory.