dhcp: if "may-fail" is true, try to renew an expired lease forever
When an active dhcp connection cannot renew the lease as the dhcp server
doen not respond, we repawn the dhclient for 3 times. If we have no luck
what happens depends upon the "may-fail" flag.
If it is false, the connection will be deactivated. If the connection
has autoconnect enable, the autoconnect will be immediately triggered.
After having exausted the autoconnect-retries, the connection will stay
deactivated for 5 minutes. Then autoconnect will try to bring the
connection up again.
If it is true, the connection will stay up, but dhcp client will never
be respawned: so we will keep the connection up without an ip and
without trying anymore to recover it: so when the dhcp server will be
back we will not be able to get the address.

With this commit when "may-fail" is true, we will keep respawning the
dhcp client to be ready when the server will be back.

The conditions checked in dhcpX_fail have been inverted: this was
required as "num_tries_left" was used as a marker to remember if we
already had an expire event (priv->dhcp4.num_tries_left ==
DHCP_NUM_TRIES_MAX).
The underling assumption that allows this change of condition order
is that when we get a timeout dhcp event and we are in "ipx_state" ==
IP_DONE we are coming from a failure to renew a lease.
Inverting the checks allows also to simplify conditions a bit.